MINOR: h3: handle errors on HEADERS parsing/QPACK decoding
Emit a CONNECTION_CLOSE if HEADERS parsing function returns an error.
This is useful to remove previous ABORT_NOW guards.

For the moment, the whole connection is closed. In the future, it may be
justified to only reset the faulting stream in some cases. This requires
the implementation of RESET_STREAM emission.

(cherry picked from commit 2bc47863ecee9b713c0b7712ae416dde6abcbe6a)
 [ad: adjusted context. H3 layer still used ncbuf instead of a plain
 buf.]
Signed-off-by: Amaury Denoyelle <adenoyelle@haproxy.com>