MINOR: ncbuf: implement ncb_is_fragmented()
Implement a new status function for ncbuf. It allows to quickly report
if a buffer contains data in a fragmented way, i.e. with gaps in between
or at start of the buffer.

To summarize, a buffer is considered as non-fragmented in the following
cases :
- a null or empty buffer
- a full buffer
- a buffer containing exactly one data block at the beginning, following
  by a gap until the end.

(cherry picked from commit e0a92a7e56b47af8586b303e33eb5ae413a51a9d)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>