MINOR: mux-quic: emit FINAL_SIZE_ERROR on invalid STREAM size
Add a check on stream size when the stream is in state Size Known. In
this case, a STREAM frame cannot change the stream size. If this is not
respected, a CONNECTION_CLOSE with FINAL_SIZE_ERROR will be emitted as
specified in the RFC 9000.

(cherry picked from commit bf91e3922bdea9034e6d19b2913d89707ae1d1f3)
Signed-off-by: Willy Tarreau <w@1wt.eu>