MINOR: mux-quic: implement accessor for sedesc
Implement a function <qcs_sc> to easily access to the stconn associated
with a QCS. This takes care of qcs.sd which may be NULL, for example for
unidirectional streams.

It is expected that in the future when implementing
STOP_SENDING/RESET_STREAM, stconn must be notify about the event. This
accessor will allow to easily test if the stconn is instantiated or not.

(cherry picked from commit 3abeb579093c5c9c40b5e95d1ff77a11738fed30)
Signed-off-by: Willy Tarreau <w@1wt.eu>