MINOR: http: Add function to get port part of a host
http_get_host_port() function can be used to get the port part of a host. It
will be used to get the port of an uri authority or a host header
value. This function only look for a port starting from the end of the
host. It is the caller responsibility to call it with a valid host value. An
indirect string is returned.

(cherry picked from commit 658f971621839f3b928da099dfe3092b47cbc958)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit 70bdf7042e86c7356cb34461148eee97c1fae440)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>