MINOR: http: Add function to detect default port
http_is_default_port() can be used to test if a port is a default HTTP/HTTPS
port. A scheme may be specified. In this case, it is used to detect defaults
ports, 80 for "http://" and 443 for "https://". Otherwise, with no scheme, both
are considered as default ports.

(cherry picked from commit ca7218aaf073627b665459bd881b2b35a481602a)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>