BUG/MINOR: peers/config: always fill the bind_conf's argument
Some generic frontend errors mention the bind_conf by its name as
"bind '%s'", but if this is used on peers "bind" lines it shows
"(null)" because the argument is set to NULL in the call to
bind_conf_uniq_alloc() instead of passing the argument. Fortunately
that's trivial to fix.

This may be backported to older versions.

(cherry picked from commit 8d158132bde98e8dfacb54d685c33eb2a982a26c)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>