BUG/MINOR: http-act: Properly generate 103 responses when several rules are used
When several "early-hint" rules are used, we try, as far as possible, to
merge links into the same 103-early-hints response. However, it only works
if there is no ACLs. If a "early-hint" rule is not executed an invalid
response is generated. the EOH block or the start-line may be missing,
depending on the rule order.

To fix the bug, we use the transaction status code. It is unused at this
stage. Thus, it is set to 103 when a 103-early-hints response is in
progress. And it is reset when the response is forwarded. In addition, the
response is forwarded if the next rule is an "early-hint" rule with an
ACL. This way, the response is always valid.

This patch must be backported as far as 2.2.

(cherry picked from commit 4c3d3d2a685bb4f33c7dae44a6f56fcce1689dcf)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit 2d9ac87c94b0469af40b2c2126db52797b4ba5bc)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit 6f67262e33dcade272a27ecd81d57d8770dbaafd)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit c68a7faae524fe9edee78a87e7666a237466e614)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>