MEDIUM: fd: make fd_insert/fd_delete atomically update fd.tgid
These functions need to set/reset the FD's tgid but when they're called
there may still be wakeups on other threads that discover late updates
and have to touch the tgid at the same time. As such, it is not possible
to just read/write the tgid there. It must only be done using operations
that are compatible with what other threads may be doing.

As we're using inc/dec on the refcount, it's safe to AND the area to zero
the lower part when resetting the value. However, in order to set the
value, there's no other choice but fd_claim_tgid() which will assign it
only if possible (via a CAS). This is convenient in the end because it
protects the FD's masks from being modified by late threads, so while
we hold this refcount we can safely reset the thread_mask and a few other
elements. A debug test for non-null masks was added to fd_insert() as it
must not be possible to face this situation thanks to the protection
offered by the tgid.

(cherry picked from commit ddedc1662487600d5124cb6d4972396438b9953c)
[wt: this is necessary to fix a design bug introduced in 2.5; ctx
 adjustment for s/tgid/1 and the fact that we don't enforce the mask
 in fd_insert() in 2.6 and older. Note that this patch had a bug with
 newstate not being used for fd.state, and needs commit 7e94b40a]
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit aa4b8a5e2ae484b92f5587eefd927fca4ba065a2)
Signed-off-by: Willy Tarreau <w@1wt.eu>