BUG/MINOR: quic: fix closing state on NO_ERROR code sent
Reception is disabled as soon as a CONNECTION_CLOSE emission is
required. An early return is done on qc_lstnr_pkt_rcv() to implement
this.

This condition is not functional if the error code sent is NO_ERROR
(0x00). To fix this, check the quic-conn flags instead of the
error code.

Currently this bug has no impact has NO_ERROR emission is not used.

This can be backported up to 2.6.

(cherry picked from commit 72d86509f1f35d1a9d2bef2a6e6b34430bd8bec6)
 [ad: adjusted context.]
Signed-off-by: Amaury Denoyelle <adenoyelle@haproxy.com>