BUG/MEDIUM: cli/threads: make "show threads" more robust on applets
Running several concurrent "show threads" in loops might occasionally
cause a segfault when trying to retrieve the stream from appctx_sc()
which may be null while the applet is finishing. It's not easy to
reproduce, it requires 3-5 sessions in parallel for about a minute
or so. The appctx_sc must be checked before passing it to sc_strm().

This must be backported to 2.6 which also has the bug.

(cherry picked from commit 52f238d32601b513ef4716c55c01b5dd0afc548a)
Signed-off-by: Amaury Denoyelle <adenoyelle@haproxy.com>