MINOR: quic: Stop looking for packet loss asap
As the TX packets are ordered by their packet number and always sent
in the same order. their TX timestamps are inspected from the older to
the newer values when we look for the packet loss. So we can stop
this search as soon as we found the first packet which has not been lost.

Must be backported to 2.6

(cherry picked from commit dc591cd6cb8cf5c0e07f720858a1f9ea62be844e)
Signed-off-by: Willy Tarreau <w@1wt.eu>