BUG/MINOR: backend: Fallback on RR algo if balance on source is impossible
If the loadbalancing is performed on the source IP address, an internal
error was returned on error. So for an applet on the client side (for
instance an SPOE applet) or for a client connected to a unix socket, an
internal error is returned.

However, when other LB algos fail, a fallback on round-robin is
performed. There is no reson to not do the same here.

This patch should fix the issue #1797. It must be backported to all
supported versions.

(cherry picked from commit 38c53944cb3306e36f6d60764c15adb206aedfb3)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit 7cac9f7ac6fbf78cb59d3a1a4e6e3ea8a37013cb)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>