Revert "BUG/MINOR: peers: set the proxy's name to the peers section name"
This reverts commit 356866accefd16458f0e3c335d1b784e24e86d2d.

It seems that an undocumented expectation of peers is based on the peers
proxy name to determine if the local peer is fully configured or not. Thus
because of the commit above, we are no longer able to detect incomplete
peers sections.

On side effect of this bug is a segfault when HAProxy is stopped/reloaded if
we try to perform a local resync on a mis-configured local peer. So waiting
for a better solution, the patch is reverted.

This patch must be backported as far as 2.5.

(cherry picked from commit 244331f6e7f2f22275cc20eeba2a0c36e27f4e7a)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>