MINOR: quic: Send packets as much as possible from qc_send_app_pkts()
Add a loop into this function to send more packets from this function
which is called by the mux. It is broken when we could not prepare
packet with qc_prep_app_pkts() due to missing available room in the
buffer used to send packets. This improves the throughput.

Must be backported to 2.6.

(cherry picked from commit dc07751ed7ebad10f49081d28a9a5ae785f53d76)
Signed-off-by: Willy Tarreau <w@1wt.eu>