MEDIUM: resolvers: continue startup if network is unavailable
When haproxy starts with a resolver section, and there is a default one
since 2.6 which use /etc/resolv.conf, it tries to do a connect() with the UDP
socket in order to check if the routes of the system allows to reach the
server.

This check is too much restrictive as it won't prevent any runtime
failure.

Relax the check by making it a warning instead of a fatal alert.

This must be backported in 2.6.

(cherry picked from commit c31577f32e0094010516a923e9f5e8cc23c23d43)
Signed-off-by: Willy Tarreau <w@1wt.eu>