MINOR: peers: Use a dedicated reconnect timeout when stopping the local peer
When a process is stopped or reload, a dedicated reconnect timeout is now
used. For now, this timeout is not used because the current code retries
immediately to reconnect to perform the local synchronization with the new
local peer, if any.

This patch is required to fix the issue #1799. It should be backported as
far as 2.0 with next fixes.

(cherry picked from commit ab4b094055fee113e91149426035c9ff7b35b2b0)
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit 810e8f2fcd514ac9c62575e0b0c1fca70b503918)
Signed-off-by: Willy Tarreau <w@1wt.eu>